BLO-351 move helper components back into the main qr scanner component file
