--- apps/admin-portal/src/app/_components/navbar/pickup-qr-scanner/loading.tsx
@@ -1,13 +1,0 @@
-import { AlertCircle } from "lucide-react";
-
-import { Alert, AlertDescription } from "@ebox/ui/alert";
-
-// Loading State Component
-export function LoadingState() {
-  return (
-    <Alert>
-      <AlertCircle className="h-4 w-4" />
-      <AlertDescription>Looking up package information...</AlertDescription>
-    </Alert>
-  );
-}
