+++ apps/admin-portal/src/app/_components/navbar/pickup-qr-scanner/scan-qr-button.tsx
@@ -5,9 +5,9 @@ import { QrCode } from "lucide-react";
 
 import { Button } from "@ebox/ui/button";
 
-import { PickupQRScanner } from "./scanner";
+import PickupQRScanner from "./scanner";
 
-export function PickupQRScannerButton() {
+export default function PickupQRScannerButton() {
   const [isQRScannerOpen, setIsQRScannerOpen] = useState(false);
 
   return (
